Refactor differences between JPN and US/PAL
Most differences stem from the event RUN_START_2 and the logic
surrounding it, which is not present in the japanese version. This
causes several variables to be assigned to values that are off by one. I
decided to fix this using a Version-dependant enum EVENTLENGTH. I'm not
sure how much this impacts readablility. Ideally, I can use descriptive
enums for all variable assignments, but for now, seeing EVENTNAME in an
assignment means the assigned value varies by 1 between versions.